+++ Changelog.rst
@@ -2,8 +2,8 @@ Version NEXTVERSION
 
 **2026-01-??**
 
-* Handle the more CF cell methods convention for anomalies
-  (https://github.com/NCAS-CMS/cfdm/issues/3??)
+* Handle cell methods that apply to norm ancillary variables
+  (https://github.com/NCAS-CMS/cfdm/issues/374)
 * Handle the more general CF cell methods convention for statistics
   applying to portions of cells
   (https://github.com/NCAS-CMS/cfdm/issues/369)